测试问题发现

+++ src/main/java/my/hive_back/module/tenant/model/entity/TenantAttendanceRule.java
@@ -26,7 +26,7 @@ public class TenantAttendanceRule {
 
     private String address;
 
-    private Integer radius;
+    private Double radius;
 
     private LocalTime workStartTime;
